clean up vecbook matching more

+++ src/simple_order.rs
@@ -2,15 +2,15 @@ use crate::order::Order;
 
 #[derive(Clone, Debug, PartialEq, Eq, Copy)]
 pub struct SimpleOrder {
-    order_id: u64,
-    quantity: u64,
-    price: u64,
+    order_id: u32,
+    quantity: u32,
+    price: u32,
 }
 
 impl SimpleOrder {
     #[must_use]
     #[allow(dead_code)]
-    pub const fn new(order_id: u64, quantity: u64, price: u64) -> Self {
+    pub const fn new(order_id: u32, quantity: u32, price: u32) -> Self {
         Self {
             order_id,
             quantity,
@@ -20,9 +20,9 @@ impl SimpleOrder {
 }
 
 impl Order for SimpleOrder {
-    type OrderId = u64;
-    type Quantity = u64;
-    type Price = u64;
+    type OrderId = u32;
+    type Quantity = u32;
+    type Price = u32;
 
     fn id(&self) -> Self::OrderId {
         self.order_id
